esg-metrics-extractor/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.49.0
pillow>=9.0.0
numpy>=1.23.0
scikit-learn>=1.0.0
psutil>=5.9.0
pymupdf>=1.21.0
huggingface_hub>=0.16.0
